Added a dummy snprintf to rig_debug to allow gcc to check format strings This found tons of errors in rig_debug statements So this patch cleans up all the files that were producing warnings or errors This should fix a few segfaults when running with debug turned on

+++ rs/gp2000.c
@@ -161,7 +161,7 @@ gp2000_set_mode(RIG *rig, vfo_t vfo, rmode_t mode, pbwidth_t width)
     int len, retval;
 
     rig_debug(RIG_DEBUG_VERBOSE, "%s: vfo=%s, mode=%s, width=%d\n", __func__,
-              rig_strvfo(vfo), rig_strvfo(mode), width);
+              rig_strvfo(vfo), rig_strvfo(mode), (int)width);
 
     switch (mode)
     {
